+++ src/authzee/compute/in_process_compute.py
@@ -8,7 +8,7 @@ __all__ = [
 ]
 
 from asyncio import as_completed, create_task, gather, sleep, Task
-from typing import Any, Callable, Dict, Iterable, List, Type
+from typing import Any, Callable, Type
 
 import jsonschema_rs
 
@@ -47,30 +47,11 @@ from authzee.types.config import (
 class InProcessCompute(ComputeModule):
 
 
-    def __init__(self):
-        super().__init__()
-        self._bg_cancel_tasks: set[Task] = set()
-
-
-    def _bg_cancel(self, t: Task | Iterable[Task]) -> None:
-        if isinstance(t, Task) is True:
-            t = [t]
-
-        for task in t:
-            task.done()
-            if task.done() is True:
-                continue
-
-            self._bg_cancel_tasks.add(task)
-            task.add_done_callback(self._bg_cancel_tasks.discard)
-            task.cancel()
-        
-
     async def start(
         self,
         execute: Callable[[str, Any], Any],
         storage_type: Type[StorageModule],
-        storage_kwargs: Dict[str, Any],
+        storage_kwargs: dict[str, Any],
         config: ComputeStartConfig
     ) -> GenericResult:
         await super().start(
@@ -91,11 +72,6 @@ class InProcessCompute(ComputeModule):
 
     async def shutdown(self, config: ComputeShutdownConfig) -> GenericResult:
         await self._storage.shutdown(config['storage'])
-        while True:
-            if len(self._bg_cancel_tasks) > 0:
-                await sleep(1)
-            else:
-                break
 
         return {
             "error": None
@@ -155,12 +131,11 @@ class InProcessCompute(ComputeModule):
         if result['error'] is not None:
             return result
 
-        all_tasks: set[Task] = set()
         context_def: ContextDef = None
         cd_page_ref = None
         cd_task: Task = None
         cd_stop = False
-        id_lookup: Dict[str, IdentityDef] = {id_type: None for id_type in request['identities']}
+        id_lookup: dict[str, IdentityDef] = {id_type: None for id_type in request['identities']}
         id_page_ref = None
         id_task: Task | set[Task] = None
         id_stop = False
@@ -182,13 +157,9 @@ class InProcessCompute(ComputeModule):
                                 config=config['list_context_defs']
                             )
                         )
-                        all_tasks.add(cd_task)
-                        cd_task.add_done_callback(all_tasks.discard)
                     else:
                         cd_page: ContextDefsPage = await cd_task
                         if cd_page['error'] is not None:
-                            self._bg_cancel_tasks(all_tasks)
-
                             return {
                                 "error": cd_page['error']
                             }
@@ -209,8 +180,6 @@ class InProcessCompute(ComputeModule):
                                     config=config['list_context_defs']
                                 )
                             )
-                            all_tasks.add(cd_task)
-                            cd_task.add_done_callback(all_tasks.discard)
 
                 else:
                     if cd_task is None:
@@ -220,16 +189,12 @@ class InProcessCompute(ComputeModule):
                                 config['get_identity_def']
                             )
                         )
-                        all_tasks.add(cd_task)
-                        cd_task.add_done_callback(all_tasks.discard)
                     else:
                         cd_result: ContextDefResult = await cd_task
                         if (
                             cd_result['error'] is not None
                             and cd_result['error']['error_type'] != "resource_not_found"
                         ):
-                            self._bg_cancel_tasks(all_tasks)
-
                             return {
                                 "error": cd_result['error']
                             }
@@ -246,8 +211,6 @@ class InProcessCompute(ComputeModule):
                                 config=config['list_identity_defs']
                             )
                         )
-                        all_tasks.add(id_task)
-                        id_task.add_done_callback(all_tasks.discard)
                     else:
                         id_page: IdentityDefsPage = await id_task
                         id_page_ref = id_page['next_page_ref']
@@ -269,8 +232,6 @@ class InProcessCompute(ComputeModule):
                                     config=config['list_identity_defs']
                                 )
                             )
-                            all_tasks.add(id_task)
-                            id_task.add_done_callback(all_tasks.discard)
 
                 else:
                     if id_task is None:
@@ -278,22 +239,18 @@ class InProcessCompute(ComputeModule):
                             create_task(self._storage.get_identity_def(it, config['get_identity_def']))
                             for it in request['identities']
                         }
-                        all_tasks.update(id_task)
-                        for t in id_task:
-                            t.add_done_callback(all_tasks.discard)
                     else:
                         for t in as_completed(id_task):
                             idr: IdentityDefResult = await t
                             if idr['error'] is not None:
                                 if idr['error']['error_type'] != "resource_not_found":
-                                    self._bg_cancel_tasks(all_tasks)
-
                                     return {
                                         "error": idr['error']
                                     }
 
                                 else:
-                                    self._bg_cancel_tasks(id_task)
+                                    # remove references to tasks so GC cleans
+                                    id_task = None
                                     break
 
                         id_stop = True
@@ -307,13 +264,9 @@ class InProcessCompute(ComputeModule):
                                 config=config['list_resource_defs']
                             )
                         )
-                        all_tasks.add(rd_task)
-                        rd_task.add_done_callback(all_tasks.discard)
                     else:
                         rd_page: ResourceDefsPage = await rd_task
                         if rd_page['error'] is not None:
-                            self._bg_cancel_tasks(all_tasks)
-
                             return {
                                 "error": rd_page['error']
                             }
@@ -334,8 +287,6 @@ class InProcessCompute(ComputeModule):
                                     config=config['list_resource_defs']
                                 )
                             )
-                            all_tasks.add(rd_task)
-                            rd_task.add_done_callback(all_tasks.discard)
 
                 else:
                     if rd_task is None:
@@ -345,16 +296,12 @@ class InProcessCompute(ComputeModule):
                                 config['get_identity_def']
                             )
                         )
-                        all_tasks.add(rd_task)
-                        rd_task.add_done_callback(all_tasks.discard)
                     else:
                         rd_result: ResourceDefResult = await rd_task
                         if (
                             rd_result['error'] is not None
                             and rd_result['error']['error_type'] != "resource_not_found"
                         ):
-                            self._bg_cancel_tasks(all_tasks)
-
                             return {
                                 "error": rd_result['error']
                             }
@@ -460,7 +407,7 @@ class InProcessCompute(ComputeModule):
                 "batch": []
             }
 
-        batch_tasks: List[Task] = []
+        batch_tasks: list[Task] = []
         for item in batch_request['batch']:
             batch_tasks.append(
                 create_task(
@@ -471,7 +418,7 @@ class InProcessCompute(ComputeModule):
                 )
             )
 
-        batch_results: List[GenericResult] = await gather(*batch_tasks)
+        batch_results: list[GenericResult] = await gather(*batch_tasks)
         batch: list = []
         for bt_result in batch_results:
             if bt_result['error'] is not None: