+++ src/authzee/jmespath.py
@@ -10,7 +10,7 @@ __all__ = [
 ]
 
 import re
-from typing import Any, Dict, List, Union
+from typing import Any
 
 
 try:
@@ -91,10 +91,10 @@ class CustomJMESPathFunctions(functions.Functions):
     )
     def _func_inner_join(
         self,
-        lhs: List[Any],
-        rhs: List[Any],
+        lhs: list[Any],
+        rhs: list[Any],
         expr: str
-    ) -> List[Dict[str, Any]]:
+    ) -> list[dict[str, Any]]:
         result = []
         for l in lhs:
             for r in rhs:
@@ -139,10 +139,10 @@ class CustomJMESPathFunctions(functions.Functions):
     )
     def _func_left_join(
         self,
-        lhs: List[Any],
-        rhs: List[Any],
+        lhs: list[Any],
+        rhs: list[Any],
         expr: str
-    ) -> List[Dict[str, Any]]:
+    ) -> list[dict[str, Any]]:
         result = []
         for l in lhs:
             lhs_match = False
@@ -193,10 +193,10 @@ class CustomJMESPathFunctions(functions.Functions):
     )
     def _func_outer_join(
         self,
-        lhs: List[Any],
-        rhs: List[Any],
+        lhs: list[Any],
+        rhs: list[Any],
         expr: str
-    ) -> List[Dict[str, Any]]:
+    ) -> list[dict[str, Any]]:
         result = []
         unmatched_rhs = set(rhs)
         for l in lhs:
@@ -275,12 +275,8 @@ class CustomJMESPathFunctions(functions.Functions):
     )
     def _func_regex_find(
         pattern: str,
-        subject: Union[str, List[str]]
-    ) -> Union[
-        None,
-        str,
-        List[Union[None, str]]
-    ]:
+        subject: str | list[str]
+    ) -> None | str | list[None | str]:
         if type(subject) is str:
             match = re.search(pattern, subject)
             if match is not None:
@@ -316,11 +312,8 @@ class CustomJMESPathFunctions(functions.Functions):
     )
     def _func_regex_find_all(
         pattern: str,
-        subject: Union[str, List[str]]
-    ) -> Union[
-        List[str],
-        List[List[str]]
-    ]:
+        subject: str | list[str]
+    ) -> list[str] | list[list[str]]:
         if type(subject) is str:
             return re.findall(pattern, subject)
 
@@ -347,17 +340,8 @@ class CustomJMESPathFunctions(functions.Functions):
     )
     def _func_regex_groups(
         pattern: str,
-        subject: Union[str, List[str]]
-    ) -> Union[
-        None,
-        List[Union[None, str]],
-        List[
-            Union[
-                None,
-                List[Union[None, str]]
-            ]
-        ]
-    ]:
+        subject: str | list[str]
+    ) -> None | list[None | str] | list[None | list[None | str]]:
         if type(subject) is str:
             match = re.search(pattern, subject)
             if match is not None:
@@ -393,11 +377,8 @@ class CustomJMESPathFunctions(functions.Functions):
     )
     def _func_regex_groups_all(
         pattern: str,
-        subject: Union[str, List[str]]
-    ) -> Union[
-        List[str],
-        List[List[str]]
-    ]:
+        subject: str | list[str]
+    ) -> list[str] | list[list[str]]:
         if type(subject) is str:
             return [list(m.groups()) if m is not None else None for m in re.finditer(pattern, subject)]